fix erroneous filenames
