+++ test/WireMock.Net.Tests/Matchers/JsonMatcherTests.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.IO;
 using FluentAssertions;
 using Newtonsoft.Json;
@@ -10,26 +9,26 @@ using Xunit;
 
 namespace WireMock.Net.Tests.Matchers
 {
-    public class JsonPartialMatcherTests
+    public class JsonMatcherTests
     {
         [Fact]
-        public void JsonPartialMatcher_GetName()
+        public void JsonMatcher_GetName()
         {
             // Assign
-            var matcher = new JsonPartialMatcher("{}");
+            var matcher = new JsonMatcher("{}");
 
             // Act
             string name = matcher.Name;
 
             // Assert
-            Check.That(name).Equals("JsonPartialMatcher");
+            Check.That(name).Equals("JsonMatcher");
         }
 
         [Fact]
-        public void JsonPartialMatcher_GetValue()
+        public void JsonMatcher_GetValue()
         {
             // Assign
-            var matcher = new JsonPartialMatcher("{}");
+            var matcher = new JsonMatcher("{}");
 
             // Act
             object value = matcher.Value;
@@ -39,30 +38,30 @@ namespace WireMock.Net.Tests.Matchers
         }
 
         [Fact]
-        public void JsonPartialMatcher_WithInvalidStringValue_Should_ThrowException()
+        public void JsonMatcher_WithInvalidStringValue_Should_ThrowException()
         {
             // Act
-            Action action = () => new JsonPartialMatcher(MatchBehaviour.AcceptOnMatch, "{ \"Id\"");
+            Action action = () => new JsonMatcher(MatchBehaviour.AcceptOnMatch, "{ \"Id\"");
 
             // Assert
             action.Should().Throw<JsonException>();
         }
 
         [Fact]
-        public void JsonPartialMatcher_WithInvalidObjectValue_Should_ThrowException()
+        public void JsonMatcher_WithInvalidObjectValue_Should_ThrowException()
         {
             // Act
-            Action action = () => new JsonPartialMatcher(MatchBehaviour.AcceptOnMatch, new MemoryStream());
+            Action action = () => new JsonMatcher(MatchBehaviour.AcceptOnMatch, new MemoryStream());
 
             // Assert
             action.Should().Throw<JsonException>();
         }
 
         [Fact]
-        public void JsonPartialMatcher_IsMatch_WithInvalidValue_And_ThrowExceptionIsFalse_Should_ReturnMismatch()
+        public void JsonMatcher_IsMatch_WithInvalidValue_And_ThrowExceptionIsFalse_Should_ReturnMismatch()
         {
             // Assign
-            var matcher = new JsonPartialMatcher("");
+            var matcher = new JsonMatcher("");
 
             // Act
             double match = matcher.IsMatch(new MemoryStream());
@@ -72,10 +71,10 @@ namespace WireMock.Net.Tests.Matchers
         }
 
         [Fact]
-        public void JsonPartialMatcher_IsMatch_WithInvalidValue_And_ThrowExceptionIsTrue_Should_ReturnMismatch()
+        public void JsonMatcher_IsMatch_WithInvalidValue_And_ThrowExceptionIsTrue_Should_ReturnMismatch()
         {
             // Assign
-            var matcher = new JsonPartialMatcher("", false, true);
+            var matcher = new JsonMatcher("", false, true);
 
             // Act
             Action action = () => matcher.IsMatch(new MemoryStream());
@@ -85,11 +84,11 @@ namespace WireMock.Net.Tests.Matchers
         }
 
         [Fact]
-        public void JsonPartialMatcher_IsMatch_ByteArray()
+        public void JsonMatcher_IsMatch_ByteArray()
         {
             // Assign
             var bytes = new byte[0];
-            var matcher = new JsonPartialMatcher("");
+            var matcher = new JsonMatcher("");
 
             // Act 
             double match = matcher.IsMatch(bytes);
@@ -99,11 +98,11 @@ namespace WireMock.Net.Tests.Matchers
         }
 
         [Fact]
-        public void JsonPartialMatcher_IsMatch_NullString()
+        public void JsonMatcher_IsMatch_NullString()
         {
             // Assign
             string s = null;
-            var matcher = new JsonPartialMatcher("");
+            var matcher = new JsonMatcher("");
 
             // Act 
             double match = matcher.IsMatch(s);
@@ -113,11 +112,11 @@ namespace WireMock.Net.Tests.Matchers
         }
 
         [Fact]
-        public void JsonPartialMatcher_IsMatch_NullObject()
+        public void JsonMatcher_IsMatch_NullObject()
         {
             // Assign
             object o = null;
-            var matcher = new JsonPartialMatcher("");
+            var matcher = new JsonMatcher("");
 
             // Act 
             double match = matcher.IsMatch(o);
@@ -127,10 +126,10 @@ namespace WireMock.Net.Tests.Matchers
         }
 
         [Fact]
-        public void JsonPartialMatcher_IsMatch_JArray()
+        public void JsonMatcher_IsMatch_JArray()
         {
             // Assign 
-            var matcher = new JsonPartialMatcher(new[] { "x", "y" });
+            var matcher = new JsonMatcher(new[] { "x", "y" });
 
             // Act 
             var jArray = new JArray
@@ -145,10 +144,10 @@ namespace WireMock.Net.Tests.Matchers
         }
 
         [Fact]
-        public void JsonPartialMatcher_IsMatch_JObject()
+        public void JsonMatcher_IsMatch_JObject()
         {
             // Assign 
-            var matcher = new JsonPartialMatcher(new { Id = 1, Name = "Test" });
+            var matcher = new JsonMatcher(new { Id = 1, Name = "Test" });
 
             // Act 
             var jobject = new JObject
@@ -163,10 +162,10 @@ namespace WireMock.Net.Tests.Matchers
         }
 
         [Fact]
-        public void JsonPartialMatcher_IsMatch_WithIgnoreCaseTrue_JObject()
+        public void JsonMatcher_IsMatch_WithIgnoreCaseTrue_JObject()
         {
             // Assign 
-            var matcher = new JsonPartialMatcher(new { id = 1, Name = "test" }, true);
+            var matcher = new JsonMatcher(new { id = 1, Name = "test" }, true);
 
             // Act 
             var jobject = new JObject
@@ -181,10 +180,10 @@ namespace WireMock.Net.Tests.Matchers
         }
 
         [Fact]
-        public void JsonPartialMatcher_IsMatch_JObjectParsed()
+        public void JsonMatcher_IsMatch_JObjectParsed()
         {
             // Assign 
-            var matcher = new JsonPartialMatcher(new { Id = 1, Name = "Test" });
+            var matcher = new JsonMatcher(new { Id = 1, Name = "Test" });
 
             // Act 
             var jobject = JObject.Parse("{ \"Id\" : 1, \"Name\" : \"Test\" }");
@@ -195,10 +194,10 @@ namespace WireMock.Net.Tests.Matchers
         }
 
         [Fact]
-        public void JsonPartialMatcher_IsMatch_WithIgnoreCaseTrue_JObjectParsed()
+        public void JsonMatcher_IsMatch_WithIgnoreCaseTrue_JObjectParsed()
         {
             // Assign 
-            var matcher = new JsonPartialMatcher(new { Id = 1, Name = "TESt" }, true);
+            var matcher = new JsonMatcher(new { Id = 1, Name = "TESt" }, true);
 
             // Act 
             var jobject = JObject.Parse("{ \"Id\" : 1, \"Name\" : \"Test\" }");
@@ -209,10 +208,10 @@ namespace WireMock.Net.Tests.Matchers
         }
 
         [Fact]
-        public void JsonPartialMatcher_IsMatch_JArrayAsString()
+        public void JsonMatcher_IsMatch_JArrayAsString()
         {
             // Assign 
-            var matcher = new JsonPartialMatcher("[ \"x\", \"y\" ]");
+            var matcher = new JsonMatcher("[ \"x\", \"y\" ]");
 
             // Act 
             var jArray = new JArray
@@ -227,10 +226,10 @@ namespace WireMock.Net.Tests.Matchers
         }
 
         [Fact]
-        public void JsonPartialMatcher_IsMatch_JObjectAsString()
+        public void JsonMatcher_IsMatch_JObjectAsString()
         {
             // Assign 
-            var matcher = new JsonPartialMatcher("{ \"Id\" : 1, \"Name\" : \"Test\" }");
+            var matcher = new JsonMatcher("{ \"Id\" : 1, \"Name\" : \"Test\" }");
 
             // Act 
             var jobject = new JObject
@@ -245,10 +244,10 @@ namespace WireMock.Net.Tests.Matchers
         }
 
         [Fact]
-        public void JsonPartialMatcher_IsMatch_WithIgnoreCaseTrue_JObjectAsString()
+        public void JsonMatcher_IsMatch_WithIgnoreCaseTrue_JObjectAsString()
         {
             // Assign 
-            var matcher = new JsonPartialMatcher("{ \"Id\" : 1, \"Name\" : \"test\" }", true);
+            var matcher = new JsonMatcher("{ \"Id\" : 1, \"Name\" : \"test\" }", true);
 
             // Act 
             var jobject = new JObject
@@ -263,10 +262,10 @@ namespace WireMock.Net.Tests.Matchers
         }
 
         [Fact]
-        public void JsonPartialMatcher_IsMatch_JObjectAsString_RejectOnMatch()
+        public void JsonMatcher_IsMatch_JObjectAsString_RejectOnMatch()
         {
             // Assign 
-            var matcher = new JsonPartialMatcher(MatchBehaviour.RejectOnMatch, "{ \"Id\" : 1, \"Name\" : \"Test\" }");
+            var matcher = new JsonMatcher(MatchBehaviour.RejectOnMatch, "{ \"Id\" : 1, \"Name\" : \"Test\" }");
 
             // Act 
             var jobject = new JObject
@@ -281,10 +280,10 @@ namespace WireMock.Net.Tests.Matchers
         }
 
         [Fact]
-        public void JsonPartialMatcher_IsMatch_JObjectWithDateTimeOffsetAsString()
+        public void JsonMatcher_IsMatch_JObjectWithDateTimeOffsetAsString()
         {
             // Assign 
-            var matcher = new JsonPartialMatcher("{ \"preferredAt\" : \"2019-11-21T10:32:53.2210009+00:00\" }");
+            var matcher = new JsonMatcher("{ \"preferredAt\" : \"2019-11-21T10:32:53.2210009+00:00\" }");
 
             // Act 
             var jobject = new JObject
@@ -295,62 +294,6 @@ namespace WireMock.Net.Tests.Matchers
 
             // Assert 
             Assert.Equal(1.0, match);
-        }
-
-        [Theory]
-        [MemberData(nameof(ValidMatches))]
-        public void JsonPartialMatcher_IsMatch_StringInputValidMatch(string value, string input)
-        {
-            // Assign
-            var matcher = new JsonPartialMatcher(value);
-
-            // Act
-            double match = matcher.IsMatch(input);
-
-            // Assert
-            Assert.Equal(1.0, match);
-        }
-
-        [Theory]
-        [MemberData(nameof(InvalidMatches))]
-        public void JsonPartialMatcher_IsMatch_StringInputWithInvalidMatch(string value, string input)
-        {
-            // Assign
-            var matcher = new JsonPartialMatcher(value);
-
-            // Act
-            double match = matcher.IsMatch(input);
-
-            // Assert
-            Assert.Equal(0.0, match);
-        }
-
-        public static IEnumerable<object[]> ValidMatches()
-        {
-            yield return new object[] { "{}", "\"test\"" };
-            yield return new object[] { "\"test\"", "\"test\"" };
-            yield return new object[] { "123", "123" };
-            yield return new object[] { "[\"test\"]", "[\"test\"]" };
-            yield return new object[] { "[\"test\"]", "[\"test\", \"test1\"]" };
-            yield return new object[] { "[123]", "[123]" };
-            yield return new object[] { "{ \"test\":\"value\" }", "{\"test\":\"value\"}" };
-            yield return new object[] { "{ \"test.test1\":\"value\" }", "{\"test\":{\"test1\":\"value\"}}" };
-            yield return new object[] { "{\"test\":{\"test1\":\"value\"}}", "{\"test\":{\"test1\":\"value\"}}" };
-            yield return new object[] { "[{ \"test.test1\":\"value\" }]", "[{\"test\":{\"test1\":\"value\"}}]" };
-        }
-
-        public static IEnumerable<object[]> InvalidMatches()
-        {
-            yield return new object[] { "\"test\"", null };
-            yield return new object[] { "\"test1\"", "\"test2\"" };
-            yield return new object[] { "123", "1234" };
-            yield return new object[] { "[\"test\"]", "[\"test1\"]" };
-            yield return new object[] { "[\"test\"]", "[\"test1\", \"test2\"]" };
-            yield return new object[] { "[123]", "[1234]" };
-            yield return new object[] { "{ \"test\":\"value\" }", "{\"test\":\"value2\"}" };
-            yield return new object[] { "{ \"test.test1\":\"value\" }", "{\"test\":{\"test1\":\"value1\"}}" };
-            yield return new object[] { "{\"test\":{\"test1\":\"value\"}}", "{\"test\":{\"test1\":\"value1\"}}" };
-            yield return new object[] { "[{ \"test.test1\":\"value\" }]", "[{\"test\":{\"test1\":\"value1\"}}]" };
         }
     }
 }
