+++ test/WireMock.Net.Tests/Matchers/JsonPartialMatcherTests.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.IO;
 using FluentAssertions;
 using Newtonsoft.Json;
@@ -9,26 +10,26 @@ using Xunit;
 
 namespace WireMock.Net.Tests.Matchers
 {
-    public class JsonMatcherTests
+    public class JsonPartialMatcherTests
     {
         [Fact]
-        public void JsonMatcher_GetName()
+        public void JsonPartialMatcher_GetName()
         {
             // Assign
-            var matcher = new JsonMatcher("{}");
+            var matcher = new JsonPartialMatcher("{}");
 
             // Act
             string name = matcher.Name;
 
             // Assert
-            Check.That(name).Equals("JsonMatcher");
+            Check.That(name).Equals("JsonPartialMatcher");
         }
 
         [Fact]
-        public void JsonMatcher_GetValue()
+        public void JsonPartialMatcher_GetValue()
         {
             // Assign
-            var matcher = new JsonMatcher("{}");
+            var matcher = new JsonPartialMatcher("{}");
 
             // Act
             object value = matcher.Value;
@@ -38,30 +39,30 @@ namespace WireMock.Net.Tests.Matchers
         }
 
         [Fact]
-        public void JsonMatcher_WithInvalidStringValue_Should_ThrowException()
+        public void JsonPartialMatcher_WithInvalidStringValue_Should_ThrowException()
         {
             // Act
-            Action action = () => new JsonMatcher(MatchBehaviour.AcceptOnMatch, "{ \"Id\"");
+            Action action = () => new JsonPartialMatcher(MatchBehaviour.AcceptOnMatch, "{ \"Id\"");
 
             // Assert
             action.Should().Throw<JsonException>();
         }
 
         [Fact]
-        public void JsonMatcher_WithInvalidObjectValue_Should_ThrowException()
+        public void JsonPartialMatcher_WithInvalidObjectValue_Should_ThrowException()
         {
             // Act
-            Action action = () => new JsonMatcher(MatchBehaviour.AcceptOnMatch, new MemoryStream());
+            Action action = () => new JsonPartialMatcher(MatchBehaviour.AcceptOnMatch, new MemoryStream());
 
             // Assert
             action.Should().Throw<JsonException>();
         }
 
         [Fact]
-        public void JsonMatcher_IsMatch_WithInvalidValue_And_ThrowExceptionIsFalse_Should_ReturnMismatch()
+        public void JsonPartialMatcher_IsMatch_WithInvalidValue_And_ThrowExceptionIsFalse_Should_ReturnMismatch()
         {
             // Assign
-            var matcher = new JsonMatcher("");
+            var matcher = new JsonPartialMatcher("");
 
             // Act
             double match = matcher.IsMatch(new MemoryStream());
@@ -71,10 +72,10 @@ namespace WireMock.Net.Tests.Matchers
         }
 
         [Fact]
-        public void JsonMatcher_IsMatch_WithInvalidValue_And_ThrowExceptionIsTrue_Should_ReturnMismatch()
+        public void JsonPartialMatcher_IsMatch_WithInvalidValue_And_ThrowExceptionIsTrue_Should_ReturnMismatch()
         {
             // Assign
-            var matcher = new JsonMatcher("", false, true);
+            var matcher = new JsonPartialMatcher("", false, true);
 
             // Act
             Action action = () => matcher.IsMatch(new MemoryStream());
@@ -84,11 +85,11 @@ namespace WireMock.Net.Tests.Matchers
         }
 
         [Fact]
-        public void JsonMatcher_IsMatch_ByteArray()
+        public void JsonPartialMatcher_IsMatch_ByteArray()
         {
             // Assign
             var bytes = new byte[0];
-            var matcher = new JsonMatcher("");
+            var matcher = new JsonPartialMatcher("");
 
             // Act 
             double match = matcher.IsMatch(bytes);
@@ -98,11 +99,11 @@ namespace WireMock.Net.Tests.Matchers
         }
 
         [Fact]
-        public void JsonMatcher_IsMatch_NullString()
+        public void JsonPartialMatcher_IsMatch_NullString()
         {
             // Assign
             string s = null;
-            var matcher = new JsonMatcher("");
+            var matcher = new JsonPartialMatcher("");
 
             // Act 
             double match = matcher.IsMatch(s);
@@ -112,11 +113,11 @@ namespace WireMock.Net.Tests.Matchers
         }
 
         [Fact]
-        public void JsonMatcher_IsMatch_NullObject()
+        public void JsonPartialMatcher_IsMatch_NullObject()
         {
             // Assign
             object o = null;
-            var matcher = new JsonMatcher("");
+            var matcher = new JsonPartialMatcher("");
 
             // Act 
             double match = matcher.IsMatch(o);
@@ -126,10 +127,10 @@ namespace WireMock.Net.Tests.Matchers
         }
 
         [Fact]
-        public void JsonMatcher_IsMatch_JArray()
+        public void JsonPartialMatcher_IsMatch_JArray()
         {
             // Assign 
-            var matcher = new JsonMatcher(new[] { "x", "y" });
+            var matcher = new JsonPartialMatcher(new[] { "x", "y" });
 
             // Act 
             var jArray = new JArray
@@ -144,10 +145,10 @@ namespace WireMock.Net.Tests.Matchers
         }
 
         [Fact]
-        public void JsonMatcher_IsMatch_JObject()
+        public void JsonPartialMatcher_IsMatch_JObject()
         {
             // Assign 
-            var matcher = new JsonMatcher(new { Id = 1, Name = "Test" });
+            var matcher = new JsonPartialMatcher(new { Id = 1, Name = "Test" });
 
             // Act 
             var jobject = new JObject
@@ -162,10 +163,10 @@ namespace WireMock.Net.Tests.Matchers
         }
 
         [Fact]
-        public void JsonMatcher_IsMatch_WithIgnoreCaseTrue_JObject()
+        public void JsonPartialMatcher_IsMatch_WithIgnoreCaseTrue_JObject()
         {
             // Assign 
-            var matcher = new JsonMatcher(new { id = 1, Name = "test" }, true);
+            var matcher = new JsonPartialMatcher(new { id = 1, Name = "test" }, true);
 
             // Act 
             var jobject = new JObject
@@ -180,10 +181,10 @@ namespace WireMock.Net.Tests.Matchers
         }
 
         [Fact]
-        public void JsonMatcher_IsMatch_JObjectParsed()
+        public void JsonPartialMatcher_IsMatch_JObjectParsed()
         {
             // Assign 
-            var matcher = new JsonMatcher(new { Id = 1, Name = "Test" });
+            var matcher = new JsonPartialMatcher(new { Id = 1, Name = "Test" });
 
             // Act 
             var jobject = JObject.Parse("{ \"Id\" : 1, \"Name\" : \"Test\" }");
@@ -194,10 +195,10 @@ namespace WireMock.Net.Tests.Matchers
         }
 
         [Fact]
-        public void JsonMatcher_IsMatch_WithIgnoreCaseTrue_JObjectParsed()
+        public void JsonPartialMatcher_IsMatch_WithIgnoreCaseTrue_JObjectParsed()
         {
             // Assign 
-            var matcher = new JsonMatcher(new { Id = 1, Name = "TESt" }, true);
+            var matcher = new JsonPartialMatcher(new { Id = 1, Name = "TESt" }, true);
 
             // Act 
             var jobject = JObject.Parse("{ \"Id\" : 1, \"Name\" : \"Test\" }");
@@ -208,10 +209,10 @@ namespace WireMock.Net.Tests.Matchers
         }
 
         [Fact]
-        public void JsonMatcher_IsMatch_JArrayAsString()
+        public void JsonPartialMatcher_IsMatch_JArrayAsString()
         {
             // Assign 
-            var matcher = new JsonMatcher("[ \"x\", \"y\" ]");
+            var matcher = new JsonPartialMatcher("[ \"x\", \"y\" ]");
 
             // Act 
             var jArray = new JArray
@@ -226,10 +227,10 @@ namespace WireMock.Net.Tests.Matchers
         }
 
         [Fact]
-        public void JsonMatcher_IsMatch_JObjectAsString()
+        public void JsonPartialMatcher_IsMatch_JObjectAsString()
         {
             // Assign 
-            var matcher = new JsonMatcher("{ \"Id\" : 1, \"Name\" : \"Test\" }");
+            var matcher = new JsonPartialMatcher("{ \"Id\" : 1, \"Name\" : \"Test\" }");
 
             // Act 
             var jobject = new JObject
@@ -244,10 +245,10 @@ namespace WireMock.Net.Tests.Matchers
         }
 
         [Fact]
-        public void JsonMatcher_IsMatch_WithIgnoreCaseTrue_JObjectAsString()
+        public void JsonPartialMatcher_IsMatch_WithIgnoreCaseTrue_JObjectAsString()
         {
             // Assign 
-            var matcher = new JsonMatcher("{ \"Id\" : 1, \"Name\" : \"test\" }", true);
+            var matcher = new JsonPartialMatcher("{ \"Id\" : 1, \"Name\" : \"test\" }", true);
 
             // Act 
             var jobject = new JObject
@@ -262,10 +263,10 @@ namespace WireMock.Net.Tests.Matchers
         }
 
         [Fact]
-        public void JsonMatcher_IsMatch_JObjectAsString_RejectOnMatch()
+        public void JsonPartialMatcher_IsMatch_JObjectAsString_RejectOnMatch()
         {
             // Assign 
-            var matcher = new JsonMatcher(MatchBehaviour.RejectOnMatch, "{ \"Id\" : 1, \"Name\" : \"Test\" }");
+            var matcher = new JsonPartialMatcher(MatchBehaviour.RejectOnMatch, "{ \"Id\" : 1, \"Name\" : \"Test\" }");
 
             // Act 
             var jobject = new JObject
@@ -280,10 +281,10 @@ namespace WireMock.Net.Tests.Matchers
         }
 
         [Fact]
-        public void JsonMatcher_IsMatch_JObjectWithDateTimeOffsetAsString()
+        public void JsonPartialMatcher_IsMatch_JObjectWithDateTimeOffsetAsString()
         {
             // Assign 
-            var matcher = new JsonMatcher("{ \"preferredAt\" : \"2019-11-21T10:32:53.2210009+00:00\" }");
+            var matcher = new JsonPartialMatcher("{ \"preferredAt\" : \"2019-11-21T10:32:53.2210009+00:00\" }");
 
             // Act 
             var jobject = new JObject
@@ -294,6 +295,62 @@ namespace WireMock.Net.Tests.Matchers
 
             // Assert 
             Assert.Equal(1.0, match);
+        }
+
+        [Theory]
+        [MemberData(nameof(ValidMatches))]
+        public void JsonPartialMatcher_IsMatch_StringInputValidMatch(string value, string input)
+        {
+            // Assign
+            var matcher = new JsonPartialMatcher(value);
+
+            // Act
+            double match = matcher.IsMatch(input);
+
+            // Assert
+            Assert.Equal(1.0, match);
+        }
+
+        [Theory]
+        [MemberData(nameof(InvalidMatches))]
+        public void JsonPartialMatcher_IsMatch_StringInputWithInvalidMatch(string value, string input)
+        {
+            // Assign
+            var matcher = new JsonPartialMatcher(value);
+
+            // Act
+            double match = matcher.IsMatch(input);
+
+            // Assert
+            Assert.Equal(0.0, match);
+        }
+
+        public static IEnumerable<object[]> ValidMatches()
+        {
+            yield return new object[] { "{}", "\"test\"" };
+            yield return new object[] { "\"test\"", "\"test\"" };
+            yield return new object[] { "123", "123" };
+            yield return new object[] { "[\"test\"]", "[\"test\"]" };
+            yield return new object[] { "[\"test\"]", "[\"test\", \"test1\"]" };
+            yield return new object[] { "[123]", "[123]" };
+            yield return new object[] { "{ \"test\":\"value\" }", "{\"test\":\"value\"}" };
+            yield return new object[] { "{ \"test.test1\":\"value\" }", "{\"test\":{\"test1\":\"value\"}}" };
+            yield return new object[] { "{\"test\":{\"test1\":\"value\"}}", "{\"test\":{\"test1\":\"value\"}}" };
+            yield return new object[] { "[{ \"test.test1\":\"value\" }]", "[{\"test\":{\"test1\":\"value\"}}]" };
+        }
+
+        public static IEnumerable<object[]> InvalidMatches()
+        {
+            yield return new object[] { "\"test\"", null };
+            yield return new object[] { "\"test1\"", "\"test2\"" };
+            yield return new object[] { "123", "1234" };
+            yield return new object[] { "[\"test\"]", "[\"test1\"]" };
+            yield return new object[] { "[\"test\"]", "[\"test1\", \"test2\"]" };
+            yield return new object[] { "[123]", "[1234]" };
+            yield return new object[] { "{ \"test\":\"value\" }", "{\"test\":\"value2\"}" };
+            yield return new object[] { "{ \"test.test1\":\"value\" }", "{\"test\":{\"test1\":\"value1\"}}" };
+            yield return new object[] { "{\"test\":{\"test1\":\"value\"}}", "{\"test\":{\"test1\":\"value1\"}}" };
+            yield return new object[] { "[{ \"test.test1\":\"value\" }]", "[{\"test\":{\"test1\":\"value1\"}}]" };
         }
     }
 }
